refactor(ssh): move parser tests out of implementation
